Update README.md
Comparing Arrays

Step 1

Examine and run the code to see how it functions. Change the value of one of the elements in the secondArray. How does the output of the program change?

Add a comment to the code each time an array is used, describing what that line of code is doing. Also add comments each time the arraysEqual variable is used, describing what that line of code is doing. You should have at least 7 comments added.

Step 2

Modify the code, now allowing the user to enter in each of the 5 elements for firstArray and secondArray from the keyboard.

Step 3

Now modify the code so that the user will first input the number of integers to be read. Assume there will be no more than 50 integers for each array.